Refactor creator handler to organizer
